accessible-nav-backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# AI模型相关
openai==1.3.0
torch==2.1.0
ultralytics==8.0.0
pillow==10.1.0

# HTTP客户端
requests==2.31.0
httpx==0.25.2

# 缓存存储
redis==5.0.1

# 工具库
python-dotenv==1.0.0

#TTS
edge-tts==6.1.9